Clean up integration with HTML now that HTML references IntersectionObserver.
Since the HTML spec has been updated to reference IntersectionObserver,
this change stops monkeypatching the HTML spec to make a change that
already exists in it.  (To some degree this is followup to #331.)
Instead, it points out where the integration exists in the HTML spec.
This avoids having two normative definitions of the same thing that
slightly disagree with each other.

It's worth noting that the monkeypatch is slightly different from what's
in HTML; where this spec's monkeypatch said "for each
IntersectionObserver whose root is in the DOMtree of that Document", the
HTML spec says "for that Document, passing in now as the timestamp".
Given the text in #run-the-update-intersection-observations-steps, the
text in HTML is correct regarding both differences, since the algorithm
both (a) expects a time and (b) builds a list of all roots in the
Document.

This also marks the definition that HTML uses as <dfn export>; this
really only serves to silence a bikeshed warning (which appears when
removing the monkeypatch) since the HTML spec isn't written using
bikeshed; however, it also means that the reference HTML makes could be
done in a spec that used bikeshed, so it deserves to be exported.